// src/types/primaryCarouselCard.tsx
export interface PrimaryCarouselCard {
    titleEn: string;
    titleAr: string;
    descriptionEn: string;
    descriptionAr: string;
    locationEn: string;
    locationAr: string;
    imageUrl: string;
    linkTextEn: string;
    linkTextAr: string;
    linkUrlEn: string;
    linkUrlAr: string;
    isPublished: boolean;
    publishedAt: Date;
    createdAt: Date;
    updatedAt: Date;
}